+++ planner/core/src/main/java/eu/seaclouds/platform/planner/core/application/decorators/ApplicationFacadeDecorator.java
@@ -14,10 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.seaclouds.platform.planner.core.decorators;
+package eu.seaclouds.platform.planner.core.application.decorators;
 
 
-import eu.seaclouds.platform.planner.core.ApplicationFacade;
+import eu.seaclouds.platform.planner.core.application.ApplicationFacade;
 
 public interface ApplicationFacadeDecorator {
