+++ planner/core/src/main/java/eu/seaclouds/platform/planner/core/application/decorators/SeaCloudsManagmentPolicyDecorator.java
@@ -14,14 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.seaclouds.platform.planner.core.decorators;
+package eu.seaclouds.platform.planner.core.application.decorators;
 
 import com.google.common.collect.ImmutableList;
 import eu.seaclouds.monitor.monitoringdamgenerator.MonitoringInfo;
-import eu.seaclouds.platform.planner.core.ApplicationFacade;
 import eu.seaclouds.platform.planner.core.DamGenerator;
 import eu.seaclouds.platform.planner.core.DamGeneratorConfigBag;
-import eu.seaclouds.platform.planner.core.agreements.AgreementGenerator;
+import eu.seaclouds.platform.planner.core.application.ApplicationFacade;
+import eu.seaclouds.platform.planner.core.application.agreements.AgreementGenerator;
 import it.polimi.tower4clouds.rules.MonitoringRules;
 import org.apache.brooklyn.util.collections.MutableMap;
 import org.apache.commons.codec.binary.Base64;
