+++ planner/core/src/main/java/eu/seaclouds/platform/planner/core/application/decorators/SlaInformationDecorator.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.seaclouds.platform.planner.core.decorators;
+package eu.seaclouds.platform.planner.core.application.decorators;
 
-import eu.seaclouds.platform.planner.core.ApplicationFacade;
 import eu.seaclouds.platform.planner.core.DamGenerator;
-import eu.seaclouds.platform.planner.core.agreements.AgreementGenerator;
+import eu.seaclouds.platform.planner.core.application.ApplicationFacade;
+import eu.seaclouds.platform.planner.core.application.agreements.AgreementGenerator;
 
 import java.util.ArrayList;
 import java.util.Arrays;
