+++ planner/core/src/main/java/eu/seaclouds/platform/planner/core/application/topology/nodetemplate/NodeTemplate.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.seaclouds.platform.planner.core.template;
+package eu.seaclouds.platform.planner.core.application.topology.nodetemplate;
 
 import java.util.List;
 import java.util.Map;
@@ -38,5 +38,5 @@ public interface NodeTemplate {
     public String getHostNodeName();
 
     List<Map<String, Object>> getArtifacts();
-    
+
 }
