+++ planner/core/src/main/java/eu/seaclouds/platform/planner/core/application/topology/nodetemplate/host/PlatformNodeTemplate.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.seaclouds.platform.planner.core.template.host;
+package eu.seaclouds.platform.planner.core.application.topology.nodetemplate.host;
 
 import com.google.common.collect.ImmutableList;
 import org.apache.brooklyn.util.collections.MutableMap;
@@ -45,7 +45,7 @@ public class PlatformNodeTemplate extends AbstractHostNodeTemplate {
         return createSimpleLocationPolicy();
     }
 
-    private Map<String, Object> createSimpleLocationPolicy(){
+    private Map<String, Object> createSimpleLocationPolicy() {
         Map<String, Object> locationPolicyDescription = MutableMap.of();
         locationPolicyDescription.put(BROOKLYN_LOCATION, "cloudfoundry-instance");
         return locationPolicyDescription;
