--- planner/core/src/main/java/eu/seaclouds/platform/planner/core/template/modifiers/NodeTemplateFacadeModifier.java
@@ -1,32 +1,0 @@
-/**
- * Copyright 2014 SeaClouds
- * Contact: SeaClouds
- * <p/>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p/>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p/>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package eu.seaclouds.platform.planner.core.template.modifiers;
-
-import eu.seaclouds.platform.planner.core.template.NodeTemplate;
-
-/**
- * A NodeTemplateFacadeModifier applies a set of changes to an NodeTemplate according to Deployer
- * reqirements
- */
-public interface NodeTemplateFacadeModifier {
-
-    void apply(NodeTemplate entitySpec);
-
-}
-
-
-
